+++ src/logic.jl
@@ -94,15 +94,15 @@ end
 end
 
 @inline function fill_order!(
-    acc::Account{TTime,OData,IData,CData},
-    order::Order{TTime,OData,IData},
+    acc::Account{TTime},
+    order::Order{TTime},
     dt::TTime,
     fill_price::Price
     ;
     fill_qty::Quantity=0.0,      # fill quantity, if not provided, order quantity is used (complete fill)
     commission::Price=0.0,       # fixed commission in quote (local) currency
     commission_pct::Price=0.0,   # percentage commission of nominal order value, e.g. 0.001 = 0.1%
-)::Trade{TTime,OData,IData} where {TTime<:Dates.AbstractTime,OData,IData,CData}
+)::Trade{TTime} where {TTime<:Dates.AbstractTime}
     inst = order.inst
     # get quote asset index
     quote_cash_index = inst.quote_cash_index