+++ src/order.jl
@@ -1,27 +1,25 @@
 using Dates
 
-mutable struct Order{TTime<:Dates.AbstractTime,OData,IData}
+mutable struct Order{TTime<:Dates.AbstractTime}
     const oid::Int
-    const inst::Instrument{IData}
+    const inst::Instrument
     const date::TTime
     const price::Price
     const quantity::Quantity   # negative = short selling
     take_profit::Price
     stop_loss::Price
-    metadata::OData
 
-    function Order(
+    function Order{TTime}(
         oid,
-        inst::Instrument{IData},
+        inst::Instrument,
         date::TTime,
         price::Price,
         quantity::Quantity
         ;
         take_profit::Price=Price(NaN),
         stop_loss::Price=Price(NaN),
-        metadata::OData=nothing
-    ) where {TTime<:Dates.AbstractTime,OData,IData}
-        new{TTime,OData,IData}(
+    ) where {TTime<:Dates.AbstractTime}
+        new{TTime}(
             oid,
             inst,
             date,
@@ -29,16 +27,37 @@ mutable struct Order{TTime<:Dates.AbstractTime,OData,IData}
             quantity,
             take_profit,
             stop_loss,
-            metadata,
         )
     end
+end
+
+function Order(
+    oid,
+    inst::Instrument,
+    date::TTime,
+    price::Price,
+    quantity::Quantity
+    ;
+    take_profit::Price=Price(NaN),
+    stop_loss::Price=Price(NaN),
+) where {TTime<:Dates.AbstractTime}
+    Order{TTime}(
+        oid,
+        inst,
+        date,
+        price,
+        quantity
+        ;
+        take_profit=take_profit,
+        stop_loss=stop_loss,
+    )
 end
 
 @inline symbol(order::Order) = symbol(order.inst)
 @inline trade_dir(order::Order) = trade_dir(order.quantity)
 @inline nominal_value(order::Order) = abs(order.quantity) * order.price * order.inst.multiplier
 
-function Base.show(io::IO, o::Order{TTime,O,I}) where {TTime,O,I}
+function Base.show(io::IO, o::Order{TTime}) where {TTime}
     date_formatter = x -> Dates.format(x, "yyyy-mm-dd HH:MM:SS")
     tp_str = isnan(o.take_profit) ? "—" : "$(format_quote(o.inst, o.take_profit)) $(o.inst.quote_symbol)"
     sl_str = isnan(o.stop_loss) ? "—" : "$(format_quote(o.inst, o.stop_loss)) $(o.inst.quote_symbol)"
@@ -50,4 +69,4 @@ function Base.show(io::IO, o::Order{TTime,O,I}) where {TTime,O,I}
               "sl=$(sl_str)")
 end
 
-Base.show(order::Order{TTime,O,I}) where {TTime,O,I} = Base.show(stdout, order)
+Base.show(order::Order{TTime}) where {TTime} = Base.show(stdout, order)